+++ s4e-backend/src/main/java/pl/cyfronet/s4e/data/repository/SceneRepositoryImpl.java
@@ -17,8 +17,8 @@ import java.util.Map;
  * Compiled from search on imgw instance of copernicus
  * limit [default 25, max = 1000]
  * offset [defaut 25]
- * orderBy [ASC - DESC]
- * sortBy [id - ingestion_time - sensing_time]
+ * order [ASC - DESC]
+ * sortBy [id - ingestionTime - sensingTime]
  * timeZone -> ZoneParameter [client time]
  * sensingFrom
  * sensingTo
@@ -56,11 +56,12 @@ public class SceneRepositoryImpl implements SceneRepositoryExt {
                 .product(Product.builder()
                         .id(rs.getLong(Scene.COLUMN_PRODUCT_ID))
                         .build())
+                .timestamp(rs.getTimestamp(Scene.COLUMN_TIMESTAMP).toLocalDateTime())
                 .build());
     }
 
     private PreparedStatement prepareQuery(Map<String, Object> params) throws SQLException {
-        PreparedStatement ps = preparedStatementBuilder.preparedStatement(connection,params);
+        PreparedStatement ps = preparedStatementBuilder.preparedStatement(connection, params);
         log.trace(ps.toString());
         return ps;
     }