+++ s4e-backend/src/test/java/pl/cyfronet/s4e/data/repository/query/QueryBuilderTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import pl.cyfronet.s4e.BasicTest;
 
-import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -30,12 +29,12 @@ public class QueryBuilderTest {
         params.put("polarisation", "Dual VV/VH");
         params.put("productType", "GRDH");
         params.put("satellitePlatform", "Sentinel-1A");
-        params.put("sensingFrom", LocalDate.of(2019, 11, 9));
-        params.put("sensingTo", LocalDate.of(2019, 11, 12));
-        params.put("ingestionFrom", LocalDate.of(2019, 11, 9));
-        params.put("ingestionTo", LocalDate.of(2019, 11, 12));
+        params.put("sensingFrom", "2019-11-09T00:00:00.000000+00:00");
+        params.put("sensingTo", "2019-11-12T00:00:00.000000+00:00");
+        params.put("ingestionFrom", "2019-11-09T00:00:00.000000+00:00");
+        params.put("ingestionTo", "2019-11-12T00:00:00.000000+00:00");
         StringBuilder resultQuery = new StringBuilder();
-        String query = "SELECT id,product_id  FROM Scene WHERE true  " +
+        String query = "SELECT id,product_id,timestamp  FROM Scene WHERE true  " +
                 "AND to_timestamp(metadata_content->>'sensing_time', 'YYYY-MM-DD\"T\"HH24:MI:SS\"Z\"') >= ?  " +
                 "AND to_timestamp(metadata_content->>'sensing_time', 'YYYY-MM-DD\"T\"HH24:MI:SS\"Z\"') <= ?  " +
                 "AND to_timestamp(metadata_content->>'ingestion_time', 'YYYY-MM-DD\"T\"HH24:MI:SS\"Z\"') >= ?  " +
@@ -43,7 +42,7 @@ public class QueryBuilderTest {
                 "AND metadata_content->>'spacecraft' = ?  AND metadata_content->>'product_type' = ?  " +
                 "AND metadata_content->>'polarisation' = ?  AND metadata_content->>'processing_level' = ?  " +
                 "AND metadata_content->>'relative_orbit_number' = ?  " +
-                "AND (metadata_content ->> 'cloud_cover')::float <= ?  LIMIT ?  OFFSET ? ;";
+                "AND (metadata_content ->> 'cloud_cover')::float <= ?  ORDER BY id DESC LIMIT ?  OFFSET ? ;";
         queryBuilder.prepareQueryAndParameters(params, parameters, resultQuery);
         assertThat(resultQuery.toString(), is(equalTo(query)));
     }
@@ -53,7 +52,7 @@ public class QueryBuilderTest {
         List<Object> parameters = new ArrayList<>();
         Map<String, Object> params = new HashMap<>();
         StringBuilder resultQuery = new StringBuilder();
-        String query = "SELECT id,product_id  FROM Scene WHERE true  LIMIT ?  OFFSET ? ;";
+        String query = "SELECT id,product_id,timestamp  FROM Scene WHERE true  ORDER BY id DESC LIMIT ?  OFFSET ? ;";
         queryBuilder.prepareQueryAndParameters(params, parameters, resultQuery);
         assertThat(resultQuery.toString(), is(equalTo(query)));
     }